finally got nodemailer working
