+++ public/javascript/login.js
@@ -4,7 +4,6 @@ async function signupForm(event) {
   const username = document.querySelector('#username-signup').value.trim();
   const email = document.querySelector('#email-signup').value.trim();
   const password = document.querySelector('#password-signup').value.trim();
-  console.log(password);
 
   if (username && email && password) {
     const response = await fetch('/api/users', {
@@ -18,8 +17,6 @@ async function signupForm(event) {
     });
 
     if (response.ok) {
-      console.log(password);
-      nodemailerJS.sendEmailNotificaiton(email);
       document.location.replace('/');
     } else {
       alert(response.statusText);
@@ -44,7 +41,6 @@ async function loginForm(event) {
     });
 
     if (response.ok) {
-      console.log('response');
       document.location.replace('/');
     } else {
       alert(response.statusText);
